requirement.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pip==23.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
setuptools==69.0.2
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.3
zipp==3.17.0
